refactor(core): relocate types.rs tests into their submodules
After the types.rs split (#749–#757), the 22 tests still sat in a
single `mod tests` at the root of `types.rs`, importing from each of
the new submodules. Move each test cluster next to the types it
actually exercises:

- agent.rs (+4): titling defaults / overrides, full AssistantConfig
  deserialization with and without an embeddings block
- llm.rs (+9): EmbeddingConfig deserialization (4), LlmConfig with
  embeddings (3), LlmConfig::default, EmbeddingProviderKind
  lowercase serialization
- features.rs (+3): MemoryConfig indexing-interval defaults
- skills_mcp.rs (+6): McpConfig stdio/http/multi-server tests,
  McpTrustLevel default

types.rs shrinks 380 → 62 lines, now only module declarations and the
crate-internal `default_true` serde helper. `cargo test -p
assistant-core --lib` reports the same 183 passed.
