+++ crates/core/src/types/agent.rs
@@ -90,3 +90,75 @@ impl Default for AgentConfig {
 fn default_agent_id() -> String {
     "default".to_string()
 }
+
+#[cfg(test)]
+mod tests {
+    use super::AssistantConfig;
+    use crate::types::llm::{EmbeddingProviderKind, LlmProviderKind};
+
+    // -- TitlingConfig defaults / overrides ----------------------------------
+
+    #[test]
+    fn test_titling_defaults_when_block_absent() {
+        // An OrgConfig with no [titling] section must deserialize cleanly and
+        // the worker should pick up the default values.
+        let toml_str = "llm_provider = \"ollama\"";
+        let cfg: AssistantConfig = toml::from_str(toml_str).unwrap();
+        let titling = cfg.titling;
+        assert!(titling.enabled, "default: enabled = true");
+        assert_eq!(titling.min_turns, 2);
+        assert_eq!(titling.long_first_message_chars, 200);
+    }
+
+    #[test]
+    fn test_titling_explicit_block_overrides_defaults() {
+        let toml_str = r#"
+            llm_provider = "ollama"
+
+            [titling]
+            enabled = false
+            min_turns = 4
+            long_first_message_chars = 500
+        "#;
+        let cfg: AssistantConfig = toml::from_str(toml_str).unwrap();
+        let titling = cfg.titling;
+        assert!(!titling.enabled);
+        assert_eq!(titling.min_turns, 4);
+        assert_eq!(titling.long_first_message_chars, 500);
+    }
+
+    // -- AssistantConfig with embeddings section -----------------------------
+
+    #[test]
+    fn full_assistant_config_with_embeddings() {
+        let toml_str = r#"
+            [llm]
+            provider = "anthropic"
+            model = "claude-opus-4-6"
+            api_key = "sk-ant-test"
+
+            [llm.embeddings]
+            provider = "openai"
+            model = "text-embedding-3-small"
+            api_key = "sk-openai-test"
+        "#;
+        let cfg: AssistantConfig = toml::from_str(toml_str).unwrap();
+        assert_eq!(cfg.llm.provider, LlmProviderKind::Anthropic);
+        let emb = cfg.llm.embeddings.expect("embeddings should be Some");
+        assert_eq!(emb.provider, EmbeddingProviderKind::OpenAI);
+        assert_eq!(emb.model.as_deref(), Some("text-embedding-3-small"));
+        assert_eq!(emb.api_key.as_deref(), Some("sk-openai-test"));
+    }
+
+    #[test]
+    fn full_assistant_config_without_embeddings() {
+        let toml_str = r#"
+            [llm]
+            provider = "ollama"
+            model = "qwen2.5:7b"
+        "#;
+        let cfg: AssistantConfig = toml::from_str(toml_str).unwrap();
+        assert_eq!(cfg.llm.provider, LlmProviderKind::Ollama);
+        assert!(cfg.llm.embeddings.is_none());
+    }
+}
